fix(ledger): move mongo conformance test to its own file
The inline test module for the MongoDB ledger conformance suite has been extracted into a dedicated `mongo_tests.rs` file, keeping the main module free of test code while preserving the same test logic and its `#[ignore]` attribute.

Auto-committed-on: dragonfly
